Major improvements to mp4_to_mng.ps1
Major update. The script now runs recursively for all systems at once. It also removes the first 17 frames (instead of just 10) to avoid black fade in. It also removes the same period from the mp3 to keep the mng and mp3 in sync. Not only that, but it also puts the results in the proper locations.

Not to mention that it looks much nicer while running.